Checkpoint Demo
Adds a simple checkpoint mechanic. If the checkpoint collides with the player object, it will store the player's data such as x, y, and later on, other variables that are added. It will also change its sprite to show that it is active. 

Note: More implementations will be needed to accompany future features that will be added. 

Added: 
- Temporary Persistence Manager Object: Invisible object that holds the conditions of the checkpoint and handles deleting the information when the game first starts and ends. It persists through the entire game; therefore, it can store information and conditions across all the rooms. (Can be used for other features going forward if desired.)
- Room Manager: For the specific room that it's placed in, it will hold similar conditions to the temporary persistence manager object but instead will handle deleting the checkpoint information once the level is completed/ends. (Can also potentially be used later on to edit other room conditions and effects.)
- Temporary Sprites: Active checkpoints have a check, and inactive checkpoints have an x.